Revamp of Project Structure and Package Renaming Front Page Reddit Backend Added
