+++ app/src/main/java/com/muaaz/joosuf/encrypted/redditclient/MainActivity.java
@@ -8,25 +8,24 @@ package com.muaaz.joosuf.encrypted.redditclient;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import com.muaaz.joosuf.encrypted.redditclient.RSSHandler.Async.RSSToPostParser;
-import com.muaaz.joosuf.encrypted.redditclient.RSSHandler.SubReddit.Post;
-
-import java.util.ArrayList;
+import com.muaaz.joosuf.encrypted.redditclient.RSSHandler.Parsers.RSSToFrontPageParser;
 
 public class MainActivity extends AppCompatActivity {
 
     public static final String TAG = "MainActivity";
-    public static final String BASE_URL = "https://www.reddit.com/r/";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        RSSToPostParser rssToPostParser = new RSSToPostParser(MainActivity.this);
-        rssToPostParser.run();
+//        RSSToSubRedditPostParser rssToPostParser = new RSSToSubRedditPostParser(MainActivity.this);
+//        rssToPostParser.run("pics");
+//
+//        ArrayList<Post> posts = rssToPostParser.getPosts();
 
-        ArrayList<Post> posts = rssToPostParser.getPosts();
+        RSSToFrontPageParser rssToFrontPageParser = new RSSToFrontPageParser(MainActivity.this);
+        rssToFrontPageParser.run("best");
+
     }
-
 }
